feat: dynamically calculate no. of pages in UpdateScreen
Modify the `ssd1306_UpdateScreen` method to dynamically
calculate the number of required RAM pages to write for
a complete update of the display.

Section 8.7 of the SSD1306 datasheet advises its memory
is divided into 8 pages for a 64px high display. Presumably
for a 128px high display the number of pages should
be doubled, and for a 32px display, halved.